+++ documentation/example/inf/iam/main.tf
@@ -47,8 +47,9 @@ module "bedrock_iam" {
 }
 
 module "ec2_iam" {
-  source  = "path/to/modules/iam/app/ec2"
-  std_map = module.com_lib.std_map
+  source       = "path/to/modules/iam/app/ec2"
+  monitor_data = data.terraform_remote_state.monitor.outputs.data
+  std_map      = module.com_lib.std_map
 }
 
 module "ecr_iam" {